requirements.txt
PySide6>=6.4.0
qt-material>=2.14
Pillow>=9.0.0
pyinstaller>=5.7.0 